reduced order partition
